requirements.txt
streamlit-lottie 
python-dotenv
google-generativeai
cryptography
streamlit==1.35.0
requests
pytz
sqlite3  # Only for documentation; it's built into Python
torch
youtube-search-python
tmdb-python